==> notebooks/2_1. Preprocesamiento_de_Datos.ipynb <==
Combinar los datasets en uno solo mejora la eficiencia del análisis , y mantener una columna de origen para luego segmentar los resultados.

import pandas as pd

# Cargar los datasets con columna de origen
dataset='Android'
dir_path = "../data/"
android = pd.read_csv(f"{dir_path}{dataset}_2k.log_structured_to_process.csv")
android['source'] = dataset
android=android[['Content', 'Level', 'datetime', 'content_length', 'source']]
android.head(2)

dataset = 'Apache'
apache = pd.read_csv(f"{dir_path}{dataset}_2k.log_structured_to_process.csv")
apache['source'] = dataset
apache=apache[['Content', 'Level', 'datetime', 'content_length', 'source']]
apache.head(2)

dataset = 'Hadoop'
hadoop = pd.read_csv(f"{dir_path}{dataset}_2k.log_structured_to_process.csv")
hadoop['source'] = dataset
hadoop=hadoop[['Content', 'Level', 'datetime', 'content_length', 'source']]
hadoop.head(2)

dataset = 'Zookeeper'
zookeeper = pd.read_csv(f"{dir_path}{dataset}_2k.log_structured_to_process.csv")
zookeeper['source'] = dataset
zookeeper=zookeeper[['Content', 'Level', 'datetime', 'content_length', 'source']]
zookeeper.head(2)

dataset = 'BGL'
bgl = pd.read_csv(f"{dir_path}{dataset}_2k.log_structured_to_process.csv")
bgl['source'] = dataset
bgl=bgl[['Content', 'Level', 'datetime', 'content_length', 'source']]
bgl.head(2)

# Unir todos los datasets en uno solo
df_all = pd.concat([android, apache, hadoop, zookeeper, bgl], ignore_index=True)
df_all.head()

df_all.to_csv(f"{dir_path}raw_data_log_structured.csv", index=False)

# Limpieza y estandarización

df_all.shape

#Remover duplicados
df_all.drop_duplicates(inplace=True)
df_all.dropna(subset=['Content', 'Level'], inplace=True)
df_all.shape

df_all.shape

df_all.head()

set(df_all['Level'])

 # Limpieza de texto y tokenización para NLP

import re
from sklearn.feature_extraction.text import ENGLISH_STOP_WORDS

stop_words = set(ENGLISH_STOP_WORDS)

def clean_text(text):
    text = str(text).lower().replace(",", " ").replace(".", " ").replace(":", " ").replace("=", " ")
    tokens = re.findall(r'\b[a-z]+\b', text)  # solo palabras alfabéticas
    tokens = [word for word in tokens if word not in stop_words]
    return ' '.join(tokens)


test = "Hello, this455 is a test."
print(clean_text(test))


df_all['Clean_Content'] = df_all['Content'].apply(clean_text)
df_all['content_length'] = df_all['Clean_Content'].apply(lambda x: len(x.split()))

df_all.shape

df_all.head()

df_all.to_csv(f"{dir_path}clean_data_log_structured.csv", index=False)

# Estadísticas Descriptivas Generales


# Total de registros
total_registros = len(df_all)

# Conteo por nivel de severidad
niveles_totales = df_all['Level'].value_counts()




# Longitud promedio del contenido
longitud_promedio = df_all['content_length'].mean()

# Fechas mínimas y máximas
min_fecha = df_all['datetime'].min()
max_fecha = df_all['datetime'].max()
print(f"Periodo: desde {min_fecha} hasta {max_fecha}")

print(f"Total de registros: {total_registros}")
print(f"Distribución por nivel:\n{niveles_totales}")
print(f"Longitud promedio del mensaje: {longitud_promedio:.2f} palabras")

**Por dataset**



# Agrupar por fuente
grouped = df_all.groupby('source')

# Recorrer cada grupo y calcular estadísticas
for name, group in grouped:
    print(f"\n Dataset: {name}")
    print(f"- Total registros: {len(group)}")
    print("- Distribución por nivel:")
    print(group['Level'].value_counts())
    print(f"- Longitud promedio del mensaje: {group['content_length'].mean():.2f} palabras")
    print(f"- Periodo: desde {group['datetime'].min()} hasta {group['datetime'].max()}")
