==> notebooks/2_2. Separar entrenamiento y validación externa .ipynb <==
import pandas as pd
import pandas as pd
import numpy as np
import random
import os
from sklearn.model_selection import train_test_split

SEED = 42 
random.seed(SEED)
np.random.seed(SEED)
os.environ['PYTHONHASHSEED'] = str(SEED)

dir_path = "../data/"
# Cargar el archivo limpio
df = pd.read_csv(f"{dir_path}clean_data_log_structured.csv")
print(set(df["source"]))
# Verificar estructura y primeras filas
print(df.shape)
print(df.columns)
df.head()

import pandas as pd

train_list = []
test_list = []

for source in df['source'].unique():
    df_source = df[df['source'] == source]

    # Calcular la cantidad mínima de elementos en cada categoría (clase minoritaria)
    level_counts = df_source['Level'].value_counts()
    min_class = level_counts.idxmin()
    min_count = level_counts[min_class]

    # Calcular cuántos elementos extraer para test desde la clase minoritaria
    test_n = max(int(min_count * 0.25), 1)

    # Diccionario para guardar cuántos tomar de cada categoría
    test_samples_per_class = {}

    for category in df_source['Level'].unique():
        df_cat = df_source[df_source['Level'] == category]

        # Mezclar aleatoriamente
        df_cat_shuffled = df_cat.sample(frac=1, random_state=42)

        # Usar el mismo número de muestras (test_n) que se tomó de la clase minoritaria
        extract_n = min(test_n, len(df_cat_shuffled))  # evita errores si alguna clase tiene < test_n

        # Dividir
        test_split = df_cat_shuffled.iloc[:extract_n]
        train_split = df_cat_shuffled.iloc[extract_n:]

        test_list.append(test_split)
        train_list.append(train_split)

# Unir los splits
train_data = pd.concat(train_list).reset_index(drop=True)
external_test_data = pd.concat(test_list).reset_index(drop=True)


## Train data

print("Train data:", train_data.shape)

print(train_data['Level'].value_counts())

# Agrupar por fuente
grouped = train_data.groupby('source')

# Recorrer cada grupo y calcular estadísticas
for name, group in grouped:
    print(f"\n Dataset: {name}")
    print(f"- Total registros: {len(group)}")
    print("- Distribución por nivel:")
    print(group['Level'].value_counts())
    print(f"- Longitud promedio del mensaje: {group['content_length'].mean():.2f} palabras")
    print(f"- Periodo: desde {group['datetime'].min()} hasta {group['datetime'].max()}")


## External test data

print("External test data:", external_test_data.shape)

print(external_test_data['Level'].value_counts())

# Agrupar por fuente
grouped = external_test_data.groupby('source')

# Recorrer cada grupo y calcular estadísticas
for name, group in grouped:
    print(f"\n Dataset: {name}")
    print(f"- Total registros: {len(group)}")
    print("- Distribución por nivel:")
    print(group['Level'].value_counts())
    print(f"- Longitud promedio del mensaje: {group['content_length'].mean():.2f} palabras")
    print(f"- Periodo: desde {group['datetime'].min()} hasta {group['datetime'].max()}")


## Save datasets

dir_path = "../process/"
train_data.to_csv(f"{dir_path}train_data.csv", index=False)
external_test_data.to_csv(f"{dir_path}external_test_data.csv", index=False)
