==> notebooks/2_3. StratifiedKFold - 10 Folds.ipynb <==
import pandas as pd
import pandas as pd
import numpy as np
import random
import os
from sklearn.model_selection import StratifiedKFold

SEED = 42 
random.seed(SEED)
np.random.seed(SEED)
os.environ['PYTHONHASHSEED'] = str(SEED)

dir_path = "../process/"
# Cargar el archivo limpio
df = pd.read_csv(f"{dir_path}train_data.csv")
print(set(df["source"]))
# Verificar estructura y primeras filas
print(df.shape)
print(df.columns)
df.head()

print("Distribución original:")
print(df['Level'].value_counts())

min_count = df['Level'].value_counts().min()

balanced_df = (
    df.groupby('Level')
    .apply(lambda x: x.sample(n=min_count, random_state=42))
    .reset_index(drop=True)
)
min_count

# Crear folds balanceados
for fold, (train_idx, val_idx) in enumerate(skf.split(df, df['Level'])):
    train_df = df.iloc[train_idx].reset_index(drop=True)
    val_df = df.iloc[val_idx].reset_index(drop=True)
    
    min_count_train = train_df['Level'].value_counts().min()
    min_count_val = val_df['Level'].value_counts().min()

    # Balancear ambos (train y val)
    train_df_balanced = (
        train_df.groupby('Level')
        .apply(lambda x: x.sample(n=min_count_train, random_state=SEED + fold))  # distinta semilla por fold
        .reset_index(drop=True)
    )

    val_df_balanced = (
        val_df.groupby('Level')
        .apply(lambda x: x.sample(n=min_count_val, random_state=SEED + fold))
        .reset_index(drop=True)
    )

    # Guardar archivos
    train_df_balanced.to_csv(os.path.join(folds_dir, f"train_fold_{fold}.csv"), index=False)
    val_df_balanced.to_csv(os.path.join(folds_dir, f"val_fold_{fold}.csv"), index=False)

    # Mostrar distribución
    print(f"\n Fold {fold} guardado:")
    print("Train:", train_df_balanced['Level'].value_counts().to_dict(), train_df_balanced.shape[0])
    print("Val:  ", val_df_balanced['Level'].value_counts().to_dict(), val_df_balanced.shape[0])
