==> notebooks/4_1. ML_Logistic Regresion_TF-IDF.ipynb <==
import pandas as pd
import numpy as np
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import classification_report, confusion_matrix
from sklearn.preprocessing import LabelEncoder
import joblib
import random
import os
import warnings
warnings.filterwarnings('ignore')

SEED = 42 
random.seed(SEED)
np.random.seed(SEED)
os.environ['PYTHONHASHSEED'] = str(SEED)

dir_path_model = "../model/"

## TF_IDF con LR

dir_path = "../process/folds/"
label_encoder = LabelEncoder()

Check if it is balanced

train_df = pd.read_csv(f"{dir_path}train_fold_0.csv")
# Codificador de etiquetas
print("Distribución original:")
print(train_df['Level'].value_counts())

loop data in each folds

for fold in range(10):
    
    print(f"***************************\ Fold {fold}")
    file= f""#f"train_fold_{fold_N}"#
    # Cargar datos
    train_df = pd.read_csv(f"{dir_path}train_fold_{fold}.csv").fillna("")
    val_df = pd.read_csv(f"{dir_path}val_fold_{fold}.csv").fillna("")
    y_train = label_encoder.fit_transform(train_df["Level"])
    y_val = label_encoder.transform(val_df["Level"])    

    # Codificar etiquetas

    # TF-IDF
    vectorizer = TfidfVectorizer(max_features=1000)
    X_train = vectorizer.fit_transform(train_df["Clean_Content"])
    X_val = vectorizer.transform(val_df["Clean_Content"])

    # Modelo
    model = LogisticRegression(max_iter=1000)
    model.fit(X_train, y_train)
    
    # Predicciones
    y_pred = model.predict(X_val)
    
    # Matriz de confusión
    cm = confusion_matrix(y_val, y_pred)
    print("Confusion Matrix:")
    print(cm)

    # Métricas
    report = classification_report(y_val, y_pred, target_names=label_encoder.classes_, output_dict=True)
    print("Classification Report:")
    print(pd.DataFrame(report).transpose())

    # Guardar modelo y vectorizador
    joblib.dump(model, f"{dir_path_model}/fold_{fold}/tfidf_logreg_fold_{fold}.joblib")
    joblib.dump(vectorizer, f"{dir_path_model}/fold_{fold}/tfidf_vectorizer_fold_{fold}.joblib")
    joblib.dump(label_encoder, f"{dir_path_model}/fold_{fold}/tfidf_le_{fold}.joblib")
    
    # Guardar métricas
    pd.DataFrame(report).transpose().to_csv(f"{dir_path_model}/results/tfidf_report_fold_{fold}.csv")
    np.savetxt(f"{dir_path_model}/results/tfidf_cm_fold_{fold}.txt", cm, fmt='%d')
    
    #break

## Evaluar el model con conjunto de validación externa

import joblib
import pandas as pd
from sklearn.metrics import classification_report, accuracy_score

# Rutas
fold = 3
dir_path_model = "../model/"


model_path = f"{dir_path_model}/fold_{fold}/tfidf_logreg_fold_{fold}.joblib"
vectorizer_path = f"{dir_path_model}/fold_{fold}/tfidf_vectorizer_fold_{fold}.joblib"
le_path = f"{dir_path_model}/fold_{fold}/tfidf_le_{fold}.joblib"
# Cargar modelo y vectorizador
model = joblib.load(model_path)
vectorizer = joblib.load(vectorizer_path)
label_encoder = joblib.load(le_path)


dir_path_ev = "../process/"

ev_df = pd.read_csv(f"{dir_path_ev}external_test_data.csv").fillna("")
print(ev_df.shape)
ev_df.head()

X_ev = ev_df["Clean_Content"]  
y_ev = ev_df["Level"] 

# Transformar texto con el vectorizador
X_ev_new = vectorizer.transform(X_ev)

y_pred = model.predict(X_ev_new)
y_pred_labels = label_encoder.inverse_transform(y_pred)


# Transformar etiquetas reales
y_true = label_encoder.transform(y_ev)

# Calcular métricas
report_dict = classification_report(
    y_true, 
    y_pred, 
    target_names=label_encoder.classes_, 
)
print(report_dict)


accuracy = accuracy_score(y_true, y_pred)
accuracy

conf_matrix = confusion_matrix(y_true, y_pred)
conf_matrix

ev_df["Predicted_Level"] = y_pred_labels
ev_df.head()

dir_path = "../process/"
# Guardar en un archivo CSV
ev_df.to_csv(f'{dir_path}4_10external_test_data_tdidf_prediction.csv', index=False)
