==> notebooks/5_1. ML_Logistic Regresion-Word2Vect.ipynb <==
import os
import numpy as np
import pandas as pd
from gensim.models import KeyedVectors
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import confusion_matrix, classification_report
from sklearn.preprocessing import LabelEncoder
import joblib
import random

SEED = 42 
random.seed(SEED)
np.random.seed(SEED)
os.environ['PYTHONHASHSEED'] = str(SEED)

## WORD_TO_VECT con LR

# Configuración
dir_path = "../process/folds/"
dir_path_model = "../model"
glove_input_file = "../glove.6B/glove.6B.100d.txt"
word2vec_output_file = "../glove.6B/glove.6B.100d.word2vec.txt"

# Cargar modelo Word2Vec
w2v_model = KeyedVectors.load_word2vec_format(word2vec_output_file, binary=False)
embedding_dim = w2v_model.vector_size

# Función para promediar vectores Word2Vec por documento
def document_vector(doc):
    words = doc.split()
    word_vecs = [w2v_model[word] for word in words if word in w2v_model]
    if len(word_vecs) == 0:
        return np.zeros(embedding_dim)
    return np.mean(word_vecs, axis=0)


# Codificador de etiquetas
label_encoder = LabelEncoder()


# Procesamiento por fold
for fold in range(10):
    print(f"\n*************************** Fold {fold}")

    # Cargar datos
    train_df = pd.read_csv(f"{dir_path}train_fold_{fold}.csv").fillna("")
    val_df = pd.read_csv(f"{dir_path}val_fold_{fold}.csv").fillna("")

    # Codificar etiquetas
    y_train = label_encoder.fit_transform(train_df["Level"])
    y_val = label_encoder.transform(val_df["Level"])

    # Representación vectorial con Word2Vec
    X_train = np.vstack(train_df["Clean_Content"].apply(document_vector).to_numpy())
    X_val = np.vstack(val_df["Clean_Content"].apply(document_vector).to_numpy())

    # Entrenar modelo
    model = LogisticRegression(max_iter=1000)
    model.fit(X_train, y_train)

    # Predicciones
    y_pred = model.predict(X_val)

    # Matriz de confusión
    cm = confusion_matrix(y_val, y_pred)
    print("Confusion Matrix:")
    print(cm)

    # Métricas
    report = classification_report(y_val, y_pred, target_names=label_encoder.classes_, output_dict=True)
    print("Classification Report:")
    print(pd.DataFrame(report).transpose())

    # Guardar modelo y encoder
    fold_dir = f"{dir_path_model}/fold_{fold}"
    os.makedirs(fold_dir, exist_ok=True)

    joblib.dump(model, f"{fold_dir}/word2vec_logreg_fold_{fold}.joblib")
    joblib.dump(label_encoder, f"{fold_dir}/word2vec_le_{fold}.joblib")

    # Guardar métricas
    os.makedirs(f"{dir_path_model}/results", exist_ok=True)
    pd.DataFrame(report).transpose().to_csv(f"{dir_path_model}/results/word2vec_report_fold_{fold}.csv")
    np.savetxt(f"{dir_path_model}/results/word2vec_cm_fold_{fold}.txt", cm, fmt='%d')
    #break

## Evaluar el model con conjunto de validación extern

dir_path = "../process/"
#logs_df = pd.read_csv(f"{dir_path}external_test_data.csv", encoding='utf-8')
train_df = pd.read_csv(f"{dir_path}train_data.csv", encoding='utf-8')
print(set(train_df["Level"]))
print(train_df.shape)
train_df=train_df.dropna(subset=['Clean_Content'])

print(train_df.shape)
train_df.head()

test_df = pd.read_csv(f"{dir_path}external_test_data.csv", encoding='utf-8')

print(set(test_df["Level"]))
print(test_df.shape)
test_df=test_df.dropna(subset=['Clean_Content'])

print(test_df.shape)
test_df.head()

# Evaluar el model con conjunto de validación extern

import pandas as pd
import numpy as np
import joblib
from sklearn.metrics import classification_report, accuracy_score, confusion_matrix
from gensim.models import KeyedVectors

# Rutas
dir_path_model = "../model"
dir_path_ev = "../process/"
word2vec_path = "../glove.6B/glove.6B.100d.word2vec.txt"


# Cargar modelo y encoder del fold 0
fold = 4
model = joblib.load(f"{dir_path_model}/fold_{fold}/word2vec_logreg_fold_{fold}.joblib")
label_encoder = joblib.load(f"{dir_path_model}/fold_{fold}/word2vec_le_{fold}.joblib")
w2v_model = KeyedVectors.load_word2vec_format(word2vec_path, binary=False)
embedding_dim = w2v_model.vector_size

# Función para representar documento con promedio de Word2Vec
def document_vector(doc):
    words = doc.split()
    word_vecs = [w2v_model[word] for word in words if word in w2v_model]
    if len(word_vecs) == 0:
        return np.zeros(embedding_dim)
    return np.mean(word_vecs, axis=0)

# Cargar dataset externo
ev_df = pd.read_csv(f"{dir_path_ev}external_test_data.csv").fillna("")
print(ev_df.shape)
ev_df.head()

# Vectorizar textos
X_ev = np.vstack(ev_df["Clean_Content"].apply(document_vector).to_numpy())
y_ev = ev_df["Level"]
y_true = label_encoder.transform(y_ev)


# Predicciones
y_pred = model.predict(X_ev)
y_pred_labels = label_encoder.inverse_transform(y_pred)


# Agregar predicciones al dataframe
ev_df["Predicted_Level"] = y_pred_labels

# Calcular métricas
report_dict = classification_report(
    y_true, 
    y_pred, 
    target_names=label_encoder.classes_, 
)
print(report_dict)


accuracy = accuracy_score(y_true, y_pred)
accuracy

conf_matrix = confusion_matrix(y_true, y_pred)
conf_matrix

ev_df["Predicted_Level"] = y_pred_labels
ev_df.head()

dir_path = "../process/"
# Guardar en un archivo CSV
ev_df.to_csv(f'{dir_path}5_10external_test_data_word2vect_prediction.csv', index=False)
