==> notebooks/5_1. Word2Vect_Results.ipynb <==
import pandas as pd
import numpy as np
import re

dir_path = "../results/"

## Validation folds

file=  "resuls_val_fold_word2vect" #"train_data" 
res_df = pd.read_csv(f"{dir_path}{file}.csv", encoding='utf-8')
res_df.head(7)

res_df_global=res_df[res_df["group"]=='macro avg']
res_df_global.head()


print("precision", np.mean(res_df_global["precision"]), "with std ", np.std(res_df_global["precision"]))
print("recall", np.mean(res_df_global["recall"]), "with std ", np.std(res_df_global["recall"]))
print("f1-score", np.mean(res_df_global["f1-score"]), "with std ", np.std(res_df_global["f1-score"]))

top_row = res_df_global.loc[res_df_global["f1-score"].idxmax()]
top_row

res_df_global=res_df[res_df["group"]=='Advertencia']
print("precision", np.mean(res_df_global["precision"]), "with std ", np.std(res_df_global["precision"]))
print("recall", np.mean(res_df_global["recall"]), "with std ", np.std(res_df_global["recall"]))
print("f1-score", np.mean(res_df_global["f1-score"]), "with std ", np.std(res_df_global["f1-score"]))

res_df_global=res_df[res_df["group"]=='Error']
print("precision", np.mean(res_df_global["precision"]), "with std ", np.std(res_df_global["precision"]))
print("recall", np.mean(res_df_global["recall"]), "with std ", np.std(res_df_global["recall"]))
print("f1-score", np.mean(res_df_global["f1-score"]), "with std ", np.std(res_df_global["f1-score"]))

res_df_global=res_df[res_df["group"]=='Informativo']
print("precision", np.mean(res_df_global["precision"]), "with std ", np.std(res_df_global["precision"]))
print("recall", np.mean(res_df_global["recall"]), "with std ", np.std(res_df_global["recall"]))
print("f1-score", np.mean(res_df_global["f1-score"]), "with std ", np.std(res_df_global["f1-score"]))
